Moved utils.ts content to connectionUtils.ts
